% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{bufferFeedingSites}
\alias{bufferFeedingSites}
\title{Buffer feeding sites}
\usage{
bufferFeedingSites(
  feedingSites,
  feedingBuffer = 100,
  crsToSet = "WGS84",
  crsToReturn = "WGS84",
  latCol = "lat",
  longCol = "long"
)
}
\arguments{
\item{feedingSites}{a data frame or an sf object}

\item{feedingBuffer}{how much to buffer the feeding site, in m}

\item{crsToSet}{if `feedingSites` is a data frame, what CRS to pass to sf::st_set_crs() (NOT transform!). If `feedingSites` is already an sf object, `crsToSet` will be overridden by whatever the object's CRS is, unless it is NA.}

\item{crsToReturn}{either "WGS84" (default) or "m" (returns in 32636, aka UTM36)}

\item{latCol}{if `feedingSites` is a data frame, the name of the column to use for latitude}

\item{longCol}{if `feedingSites` is a data frame, the name of the column to use for longitude}
}
\value{
sf object containing feeding site polygons, buffered
}
\description{
Buffer feeding sites by a specified number of meters
}
