% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{calcSpeeds}
\alias{calcSpeeds}
\title{Calculate speeds}
\usage{
calcSpeeds(df, grpCol, longCol, latCol)
}
\arguments{
\item{df}{a data frame}

\item{grpCol}{column to group by}

\item{longCol}{column containing longitudes}

\item{latCol}{column containing latitudes}
}
\value{
A data frame with speeds added
}
\description{
Calculates speeds, an operation that needs to happen several times in Marta's data cleaning code
}
