% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{calcSpeedsVert}
\alias{calcSpeedsVert}
\title{Calculate vertical speeds (altitude)}
\usage{
calcSpeedsVert(df, grpCol, altCol, speedCol)
}
\arguments{
\item{df}{a data frame}

\item{grpCol}{column to group by}

\item{altCol}{column containing altitude values}

\item{speedCol}{column giving ground speed, so we can restrict this to flight only}
}
\value{
A data frame with speeds added
}
\description{
Calculates vertical "speeds", an operation that needs to happen several times in order to clean the altitude values
}
