% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{calculateSpeeds}
\alias{calculateSpeeds}
\title{Calculate speeds based on lead/lag}
\usage{
calculateSpeeds(dataset, idCol, timestampCol, x, y)
}
\arguments{
\item{dataset}{the dataset to remove outliers from}

\item{idCol}{name of the column containing the animal identifier}

\item{timestampCol}{name of the column containing timestamps}

\item{x}{name of the column containing longitude values}

\item{y}{name of the column containing latitude values}
}
\value{
a dataset with speeds calculated
}
\description{
Supporting function for removeSpeedOutliers
}
