% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{cleanAndMetadata}
\alias{cleanAndMetadata}
\title{Clean data and extract metadata}
\usage{
cleanAndMetadata(
  df,
  speedLower = NULL,
  speedUpper = NULL,
  dateCol = "dateOnly"
)
}
\arguments{
\item{df}{a data frame to filter. Must include column `ground_speed`.}

\item{speedLower}{a single numeric value, the lower limit for ground speed to be included (m/s)}

\item{speedUpper}{a single numeric value, the upper limit for ground speed to be included (m/s)}

\item{dateCol}{the name of the column containing dates tracked. Default is "dateOnly".}
}
\value{
A list: "filteredData" = filtered dataset; "tagsMetadata" = tag-by-tag metadata
}
\description{
Extract metadata by tag, and filter/clean the dataset for speed, gps satellites, and heading
}
