% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{computeProbs}
\alias{computeProbs}
\title{Compute probabilities of gaining and losing edges.}
\usage{
computeProbs(graphList)
}
\arguments{
\item{graphList}{a list of network graphs}
}
\value{
a long-format data frame containing probabilities of losing/adding edges
}
\description{
Given a list of graphs, compute the probability at each time step of gaining or losing edges, looking two timesteps back.
}
