% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{convert}
\alias{convert}
\title{Convert coordinates of an sf object}
\usage{
convert(obj, crsMeters = 32636)
}
\arguments{
\item{obj}{an sf object to be buffered}

\item{crsMeters}{crs with units of meters to be used. Default is 32636 (Israel, UTM zone 36)}
}
\value{
A buffered sf object
}
\description{
Given an sf object in WGS84, convert it to a CRS with meters as the units.
}
