% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{convertAndBuffer}
\alias{convertAndBuffer}
\title{Given an sf object in WGS84, convert it to a CRS with meters as the units, buffer by a given distance, and then convert it back.}
\usage{
convertAndBuffer(obj, dist = 50, crsMeters = 32636)
}
\arguments{
\item{obj}{an sf object to be buffered}

\item{dist}{buffer distance, in meters (m)}

\item{crsMeters}{crs with units of meters to be used. Default is 32636 (Israel, UTM zone 36)}
}
\value{
A buffered sf object
}
\description{
Given an sf object in WGS84, convert it to a CRS with meters as the units, buffer by a given distance, and then convert it back.
}
