% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{createDirectedMatrices}
\alias{createDirectedMatrices}
\title{Create directed matrices}
\usage{
createDirectedMatrices(
  dataset,
  distThreshold,
  sim = SimlDataPntCnt,
  co = CoocurCountr,
  latCol = "location_lat",
  longCol = "location_long"
)
}
\arguments{
\item{dataset}{the vulture dataset}

\item{distThreshold}{distance threshold for what is considered an "interaction"}

\item{sim}{# number of simulations?? not sure about this one.}

\item{co}{#  number of co-occurrences? not sure about this one either.}
}
\value{
A list: "SimlDataPntCnt" = sim, "CoocurCountr" = co
}
\description{
Create directed matrices from vulture data
}
