% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{createDirectedMatricesRevised}
\alias{createDirectedMatricesRevised}
\title{Create directed matrices (revised)}
\usage{
createDirectedMatricesRevised(
  dataset,
  distThreshold,
  latCol = "location_lat",
  longCol = "location_long",
  idCol = "trackId"
)
}
\arguments{
\item{dataset}{a time-grouped dataset to operate on. Must be a data frame.}

\item{distThreshold}{a distance threshold for interactions (m)}

\item{latCol}{character; the name of the column containing the latitude coordinates (WGS84 assumed). Default is "location_lat".}

\item{longCol}{character; the name of the column containing the longitude coordinates (WGS84 assumed). Default is "location_long".}

\item{idCol}{character; the name of the column containing unique individual identifiers. Default is "trackId".}
}
\value{
A list of edge lists (one per time group)
}
\description{
Revised version of Nitika's function. In beta.
}
