% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{downloadVultures}
\alias{downloadVultures}
\title{Data download}
\usage{
downloadVultures(
  loginObject,
  extraSensors = F,
  removeDup = T,
  dateTimeStartUTC = NULL,
  dateTimeEndUTC = NULL
)
}
\arguments{
\item{loginObject}{A Movebank login object, created by passing a username and password to move::movebankLogin}
}
\value{
A movestack
}
\description{
Download vulture project data from the Movebank repository, with some minor specifications. Note that you must specify your movebank credentials.
}
