% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{filterLocs}
\alias{filterLocs}
\title{Filter locs}
\usage{
filterLocs(df, speedThreshLower = NULL, speedThreshUpper = NULL)
}
\arguments{
\item{df}{a data frame to filter}

\item{speedThreshLower}{a single numeric value, the lower limit for ground speed to be included (m/s)}

\item{speedThreshUpper}{a single numeric value, the upper limit for ground speed to be included (m/s)}
}
\value{
A list of filtered data frames.
}
\description{
Filter dataset for reasonableness
}
