% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_roosts}
\alias{get_roosts}
\title{Get roosts}
\usage{
get_roosts(
  id,
  timestamp,
  x,
  y,
  ground_speed,
  speed_units = c("m/s", "km/h"),
  buffer = 1,
  twilight = 61,
  morning_hours = c(0:12),
  night_hours = c(13:23)
)
}
\arguments{
\item{id}{animal identifier}

\item{timestamp}{object of class `as.POSIXct`}

\item{x}{longitude, in decimal degrees}

\item{y}{latitude, in decimal degrees}

\item{ground_speed}{ground speed of the animal}

\item{speed_units}{units of speed (either "m/s" or "km/h"). If speed is provided in "km/h" it is transformed to "m/s".}

\item{buffer}{optional, numerical value indicating the number of kms to consider the buffer for the roost calculation. The pre-defined value is 1km}

\item{twilight}{optional, numerical value indicating number of minutes to consider the twilight for calculating the day and night positions. If set to 0, the night period starts at sunset and the day period starts at sunrise. The pre-defined value is 61, so the night period starts 61 minutes before sunset and the day period starts 61 minutes after sunrise}

\item{morning_hours}{optional, vector indicating the range of hours (in UTC) that are considered the morning period. The pre-defined vector is `c(0:12)`}

\item{night_hours}{optional, vector indicating the range of hours (in UTC) that are considered the night period. The pre-defined vector is `c(13:23)`}
}
\description{
With several methods, get the roost locations of a vulture on each night. This function was written by Marta Acácio. The function consecutively calculates the night roost based on the following five methods:
1.  It is the last GPS location of the day, and it is at night (and during the "night hours"), and the speed is equal or less than 4m/s (i.e., the bird was considered to not be flying);
2.  It is the last GPS location of the day, and it is at night (and during the "night hours"), and the speed is NA;
3.  It is the first GPS location of the day, and it is at night (and during the "morning hours"), and the speed is equal or less than 4m/s;
4.  It is the first GPS location of the day, and it is at night (and during the "morning hours"), and the speed is NA;
5.  The last GPS location of the day (independently of the light or the hours) is within the defined buffer (pre-defined, 1km) of the first GPS location of the following day.
The roost day is assigned in the following way:
1.  If it is the last location of the day, it is that day's night roost;
2.  If it was the first location of the day, it was the previous day's night roost;
3.  If for a particular date, the roost was calculated using more than 1 method, the selected roost is the earliest calculated roost.
}
