% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{getRoostEdges}
\alias{getRoostEdges}
\title{Create co-roosting edge list}
\usage{
getRoostEdges(
  dataset,
  mode = "distance",
  roostPolygons = NULL,
  distThreshold,
  latCol = "location_lat",
  longCol = "location_long",
  idCol = "trackId",
  dateCol = "date",
  roostCol = "roostID",
  crsToSet = "WGS84",
  return = "edges"
)
}
\arguments{
\item{dataset}{A data frame. May have one of two formats, depending on `mode`: either ID/date/lat/long (works for `mode` = "distance", or `mode` = "polygon" if `roostPolygons` is also supplied); or ID/date/roostID (work only for `mode` = "polygon"). May have additional columns.}

\item{mode}{One of "distance" (default) or "polygon". If "distance", creates a distance-based network with edges each night. If "polygon", creates a shared-polygon-based network.}

\item{roostPolygons}{blah}

\item{distThreshold}{blah}

\item{latCol}{blah}

\item{longCol}{blah}

\item{idCol}{blah}

\item{dateCol}{blah}

\item{roostCol}{blah}

\item{crsToSet}{Default WGS84.}

\item{return}{One of "edges" (default, returns an edgelist, would need to be used in conjunction with includeAllVertices = T in order to include all individuals, since otherwise they wouldn't be included in the edgelist); "sri" (returns a data frame with three columns, ID1, ID2, and sri. Includes pairs whose SRI values are 0, which means it includes all individuals and renders includeAllVertices obsolete.); and "both" (returns a list with two components: "edges" and "sri" as described above.)}
}
\description{
Function to create an edgelist based on roost locations. Two modes: either create a straightforward distance-based network based on a distance threshold (analogous to getFeedingEdges and getFlightEdges, except that there's only one point per individual per night); or create a shared-roost network based on roost polygon assignment.
}
