% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{getStats}
\alias{getStats}
\title{Get Stats}
\usage{
getStats(df, idCol = "Nili_id")
}
\arguments{
\item{df}{Any data frame with IDs}

\item{idCol}{The name of the column in the dataset containing vulture ID's. Defaults to "Nili_id" (assuming you have joined the Nili_ids from the who's who table).}
}
\value{
An edge list containing the following columns: `timegroup` gives the numeric index of the timegroup during which the interaction takes place. `minTimestamp` and `maxTimestamp` give the beginning and end times of that timegroup. `ID1` is the id of the first individual in this edge, and `ID2` is the id of the second individual in this edge.
}
\description{
This function takes in a data frame with rows representing data points and returns the number of rows, columns and
individuals in the data frame.
}
