% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{gpsTimeFilter}
\alias{gpsTimeFilter}
\title{GPS time to fix filter}
\usage{
gpsTimeFilter(dataset, maxTime = 89)
}
\arguments{
\item{dataset}{A dataset with columns names: gps_time_to_fix}

\item{maxTime}{Max time for gps time fix, default is 89}
}
\value{
A dataset with outliers removed
}
\description{
This function takes in a dataset and removes outliers with columns: gps_time_to_fix > threshold
}
