% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{headingFilter}
\alias{headingFilter}
\title{Heading Filter}
\usage{
headingFilter(dataset)
}
\arguments{
\item{dataset}{A dataset with columns names: heading}
}
\value{
A dataset with outliers removed
}
\description{
This function takes in a dataset and removes outliers with columns: heading < 0 or heading > 360
}
