% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{inMaskFilter}
\alias{inMaskFilter}
\title{In Mask Filter}
\usage{
inMaskFilter(
  dataset,
  mask,
  inMaskThreshold = 0.33,
  crs = "WGS84",
  longCol = "location_long.1",
  latCol = "location_lat.1",
  dateCol = "dateOnly",
  idCol = "Nili_id",
  reMask = T,
  quiet = T
)
}
\arguments{
\item{dataset}{A dataset with columns: longCol, latCol, dateCol, idCol}

\item{mask}{The object to use to mask the data. Passed to `vultureUtils::maskData()`. Must be an sf object.}

\item{inMaskThreshold}{Proportion of an individual's days tracked that must fall within the mask. Default is 0.33 (one third of days tracked). If a number >1 is supplied, will be interpreted as number of days that fall in the mask. Passed to `vultureUtils::mostlyInMask()`. Must be numeric.}

\item{crs}{Coordinate Reference System to check for and transform to, for both the GPS data and the mask. Default is "WGS84". This value is passed to `vultureUtils::maskData()`. Must be a valid CRS or character string coercible to CRS.}

\item{longCol}{The name of the column in the dataset containing longitude values. Defaults to "location_long.1". Passed to `vultureUtils::maskData()`.}

\item{latCol}{The name of the column in the dataset containing latitude values. Defaults to "location_lat.1". Passed to `vultureUtils::maskData()`.}

\item{dateCol}{The name of the column in the dataset containing dates. Defaults to "dateOnly". Passed to `vultureUtils::mostlyInMask()`.}

\item{idCol}{The name of the column in the dataset containing vulture ID's. Defaults to "Nili_id" (assuming you have joined the Nili_ids from the who's who table).}

\item{reMask}{Whether or not to re-mask after removing individuals that spend less than `inMaskThreshold` in the mask area. Default is T.}

\item{quiet}{Whether to silence the message that happens when doing spatial joins. Default is T.}
}
\description{
This function filters out individuals that have a percentage of points outside of the mask. The percentage is given by
inMaskThreshold. Optionally, points falling outside the mask can also be removed with reMask set to true.
Steps: 1. Using the `mask` object, get a list of the individuals in `dataset` that spend at least `inMaskThreshold`
proportion of their time inside the mask area. 2. Restrict `dataset` to only these individuals. 3. Re-apply the mask to
restrict the remaining points to those that fall within `mask`.
}
