% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{israelCRS}
\alias{israelCRS}
\title{CRS for Israel shapefile}
\format{
A crs object
\describe{
  \item{}{}
  ...
}
}
\source{
\url{}
}
\usage{
israelCRS
}
\description{
The CRS for the Israel shapefile, saved separately because it doesn't seem to load properly...
}
\keyword{datasets}
