% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{makeGIF}
\alias{makeGIF}
\title{Make a gif from a graph list}
\usage{
makeGIF(plotList, fileName, interval = 0.1)
}
\arguments{
\item{plotList}{a list of plots}

\item{fileName}{where to write the file. Must end with .gif.}

\item{interval}{a positive number to set the time interval of the animation (unit in seconds); default to 0.1.}
}
\value{
A gif file
}
\description{
Given a list of plots, create a gif.
}
