% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{makeGraph}
\alias{makeGraph}
\title{Make a single graph from an edge list.}
\usage{
makeGraph(
  mode = "edgelist",
  data,
  weighted = FALSE,
  id1Col = "ID1",
  id2Col = "ID2",
  vertices = NULL
)
}
\arguments{
\item{mode}{One of "edgelist" (default; creates a graph from an edgelist) or "sri" (creates a graph from a data frame containing each dyad once, with its SRI value).}

\item{data}{a data frame containing edges and their associated `timegroup`s. Returned from vultureUtils::get*Edges().}

\item{weighted}{whether or not the resulting graphs should have weights attached. Default is FALSE (unweighted graph).}

\item{id1Col}{name of the column in `edges` containing the ID of the first individual in a dyad. Default is "ID1".}

\item{id2Col}{name of the column in `edges` containing the ID of the second individual in a dyad. Default is "ID2".}

\item{vertices}{optional. Either NULL (default) or a data frame/vector with vertex names and optional metadata. If a vector is provided, it will be coerced to a data frame (see documentation for igraph::graph_from_data_frame).}
}
\value{
an igraph object. An undirected graph that is either weighted or unweighted, depending on whether `weighted` is T or F.
}
\description{
A wrapper of igraph::graph_from_data_frame. Takes into account the formatting returned from vultureUtils::get*Edges(). Has two modes--can either take an edgelist (optionally with a full list of vertices) or an SRI data frame.
}
