% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{makeGraphs}
\alias{makeGraphs}
\title{Make a list of networks, given a time interval}
\usage{
makeGraphs(
  edges,
  fullData,
  interval,
  dateTimeStart = NULL,
  dateTimeEnd = NULL,
  id1Col = "ID1",
  id2Col = "ID2",
  weighted = FALSE,
  dropFinalGraph = FALSE
)
}
\arguments{
\item{edges}{a data frame containing edges and their associated `timegroup`s.}

\item{fullData}{a data frame containing the `timegroup` column and `timestamp`s of vulture relocations. This will be used to classify timegroups by interval.}

\item{interval}{A character string specifying an interval such as "3 days" or "2 hours" or "1 month" or "1.5 hours" or "3 days 2 hours". Interval will be coerced to a duration object using lubridate::as.duration()}

\item{dateTimeStart}{the dateTime object that defines the beginning of the time period to be divided. If not specified, defaults to the earliest `timestamp` in `fullData`. Must be in one of the following formats: "YYYY-MM-DD hh:mm:ss" or "YYYY-MM-DD hh:mm" or "YYYY-MM-DD". Hours must use 24 hour time--e.g. 5:00 pm would be 17:00.}

\item{dateTimeEnd}{the dateTime object that defines the end of the time period to be divided. If not specified, defaults to the latest `timestamp` in `fullData`. Must be in one of the following formats: "YYYY-MM-DD hh:mm:ss" or "YYYY-MM-DD hh:mm" or "YYYY-MM-DD". Hours must use 24 hour time--e.g. 5:00 pm would be 17:00.}

\item{id1Col}{name of the column in `edges` containing the ID of the first individual in a dyad}

\item{id2Col}{name of the column in `edges` containing the ID of the second individual in a dyad}

\item{weighted}{whether or not the resulting graphs should have weights attached}
}
\value{
A list of igraph graph objects
}
\description{
Inputs: a data frame containing time-grouped edges, and given an interval for making the networks. This function separates the data frame into a list of edge lists according to the provided time interval. Then, it generates a network (weighted or unweighted) for each of the edge lists. Note that depending on whether or not `interval` divides evenly into the range between `dateTimeStart` and `dateTimeEnd`, the last graph may be created from a smaller time period than the other graphs.
}
