% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{makeNetwork_coFeeding}
\alias{makeNetwork_coFeeding}
\title{Create co-feeding edge list
# XXX to do: - make this one single wrapper that can handle feeding, flying, and roosting. - make *all* possible input parameters adjustable here, with reasonable defaults. - make sure documentation for all those parameters is up to date, including a clear indication of which other function they get passed to.}
\usage{
makeNetwork_coFeeding(
  dataset,
  mask,
  roostPolygons,
  inMaskThreshold = 0.33,
  consecThreshold = 2,
  distThreshold = 50,
  speedThreshUpper = 5,
  speedThreshLower = NULL
)
}
\arguments{
\item{dataset}{The GPS dataset to be used to create the edge list.}

\item{mask}{The object to use to mask the data. Passed to `vultureUtils::maskData()`. Must be an sf object.}

\item{roostPolygons}{Roost polygons. Must be an sf object with a CRS that matches the dataset CRS.}

\item{inMaskThreshold}{Proportion of an individual's days tracked that must fall within the mask. Default is 0.33 (one third of days tracked). Passed to `vultureUtils::mostlyInMask()`. Must be numeric.}

\item{consecThreshold}{Minimal number of co-occurrences for considering a viable pair of interacting vultures (default is 2 consecutive time steps). Passed to `vultureUtils::spaceTimeGroups()`. Must be numeric.}

\item{distThreshold}{The maximum distance (in meters) at which vultures are considered interacting. Default is 50 for co-feeding. Passed to `vultureUtils::spaceTimeGroups()`. Must be numeric.}
}
\value{
An edge list containing the following columns: `timegroup` gives the numeric index of the timegroup during which the interaction takes place. `minTimestamp` and `maxTimestamp` give the beginning and end times of that timegroup. `ID1` is the trackID of the first individual in this edge, and `ID2` is the trackID of the second individual in this edge.
}
\description{
Given a dataset of GPS points, a geographic mask (optional), and some roost polygons, create an edge list.
}
