% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mask}
\alias{mask}
\title{Shapefile in the rough shape of Israel}
\format{
An sf object with 1 feature and 2 fields, type POLYGON
\describe{
  \item{Regional_polygon}{approximate extent of the Israel area}
  ...
}
}
\source{
\url{}
}
\usage{
mask
}
\description{
A shapefile used to mask the vulture data to exclude points outside of Israel.
}
\keyword{datasets}
