% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{maskIsrael}
\alias{maskIsrael}
\title{Mask Israel}
\usage{
maskIsrael(dataset, longCol = "location_long", latCol = "location_lat", crs)
}
\arguments{
\item{dataset}{a dataset to mask}

\item{longCol}{the name of the column in the dataset containing longitude values}

\item{latCol}{the name of the column in the dataset containing latitude values}

\item{crs}{(To be passed to `st_set_crs()`). One of (i) character: a string accepted by GDAL, (ii) integer, a valid EPSG value (numeric), or (iii) an object of class crs.}
}
\value{
A movestack
}
\description{
Crop data to only include locations within Israel
}
