% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{mostlyInIsrael}
\alias{mostlyInIsrael}
\title{Which mostly in Israel?}
\usage{
mostlyInIsrael(dataset, israelDataset, thresh = 0.333, dateCol = "dateOnly")
}
\arguments{
\item{dataset}{the full dataset, before masking to Israel only}

\item{israelDataset}{the dataset after being masked to Israel (output of maskIsrael function)}

\item{thresh}{proportion (between 0 and 1) of a vulture's total tracked days that it spent in Israel}

\item{dateCol}{the name of the column containing dates (must be the same in `dataset` and `israelDataset`). Defaults to "dateOnly".}
}
\value{
A vector of trackIds for vultures
}
\description{
Only include individuals that spent a certain proportion of their total time tracked in Israel
}
