% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{plotGraphs}
\alias{plotGraphs}
\title{Create plots from a list of graphs}
\usage{
plotGraphs(graphList, coords = "fixed")
}
\arguments{
\item{graphList}{a list of graph objects.}

\item{coords}{either "fixed" (use same coordinates for all plots, to visualize change over time), or "free" (different coords for each plot)}
}
\value{
A list of plot objects
}
\description{
Given a list of graphs, make plots. Optionally, use a consistent layout to visualize change over time.
}
