% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{preciseFilter}
\alias{preciseFilter}
\title{Precise Filter}
\usage{
preciseFilter(dataset)
}
\arguments{
\item{dataset}{A dataset with columns names: gps_satellite_count, gps_hdop}
}
\value{
A dataset with outliers removed
}
\description{
This function takes in a dataset and removes outliers with columns: gps_satellite_count < 4 and gps_hdop > 5
}
