% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{removeCaptures}
\alias{removeCaptures}
\title{Remove Capture Cage periods}
\usage{
removeCaptures(
  data,
  captureSites,
  AllCarmelDates,
  distance = 500,
  idCol = "Nili_id",
  start.day = 1,
  start.month = 8,
  end.day = 30,
  end.month = 11
)
}
\arguments{
\item{captureSites}{A data frame of capture sites}

\item{AllCarmelDates}{A data frame of dates containing captures in Carmel}

\item{idCol}{Name of the column in the data containing individual IDs. Default is Nili_id. Used for getting roosts.}

\item{start.day}{When combined with `start.month`, starting date of the capture period. Default is 1.}

\item{start.month}{Starting month of the capture period. Default is August.}

\item{end.day}{When combined with `end.month`, ending date of the capture period. Default is 30.}

\item{end.month}{Ending month of the capture period. Default is November.}

\item{dataset}{A dataset}
}
\value{
A dataset with capture cage periods removed
}
\description{
This function takes in a dataset of vultures and removes data points which contain known times the individual was in a capture cage.
Note that it is important that the roosts, captureSites, and AllCarmelDates is provided after being read in as a data.frame as it is necessary for this function to work.
}
