% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{removeSpeedOutliers}
\alias{removeSpeedOutliers}
\title{Remove speed outliers}
\usage{
removeSpeedOutliers(
  dataset,
  idCol = "trackId",
  timestampCol = "timestamp",
  x = "location_long",
  y = "location_lat",
  includeSpeedCols = FALSE
)
}
\arguments{
\item{dataset}{the dataset to remove outliers from}

\item{idCol}{name of the column containing the animal identifier}

\item{timestampCol}{name of the column containing timestamps}

\item{x}{name of the column containing longitude values}

\item{y}{name of the column containing latitude values}

\item{includeSpeedCols}{logical, whether to return a data frame including the calculated speeds (T) or remove the calculated speed columns (F, default)}
}
\value{
a data frame with speed-based outliers removed.
}
\description{
This function identifies points that are errors/outliers based on the inferred lead/lag speed. Written by Marta Acácio as part of the data cleaning workflow; adapted for this package by Kaija Gahm.
}
