% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{removeUnnecessaryVars}
\alias{removeUnnecessaryVars}
\title{Remove unnecessary vars}
\usage{
removeUnnecessaryVars(dataset)
}
\arguments{
\item{dataset}{a dataset to remove variables from}
}
\value{
A dataset, with variables removed
}
\description{
Remove variables we don't need, so the data is smaller
}
