% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{satelliteFilter}
\alias{satelliteFilter}
\title{Satellite Filter}
\usage{
satelliteFilter(dataset, minSatellites = 3)
}
\arguments{
\item{dataset}{A dataset with columns names: heading}

\item{minSatellites}{Minimum number of satellites required, default is 3}
}
\value{
A dataset with outliers removed
}
\description{
This function takes in a dataset and removes outliers with columns: gps_satellite_count < minSatellites
}
