% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sliceTemporal}
\alias{sliceTemporal}
\title{Make temporal slices of an edge list}
\usage{
sliceTemporal(
  edges,
  interval,
  dateTimeStart = NULL,
  dateTimeEnd = NULL,
  id1Col = "ID1",
  id2Col = "ID2",
  wminTimestampCol = "minTimestamp",
  maxTimestampCol = "maxTimestamp"
)
}
\arguments{
\item{edges}{a data frame containing edges and their associated `timegroup`s. Returned by `vultureUtils::get*Edges()`}

\item{interval}{a character string specifying an interval such as "3 days" or "2 hours" or "1 month" or "1.5 hours" or "3 days 2 hours". Interval will be coerced to a duration object using lubridate::as.duration()}

\item{dateTimeStart}{a dateTime object that defines the beginning of the time period to be divided. If not specified, defaults to the earliest `timestamp` in `edges`. Must be in one of the following formats: "YYYY-MM-DD hh:mm:ss" or "YYYY-MM-DD hh:mm" or "YYYY-MM-DD". Hours must use 24 hour time--e.g. 5:00 pm would be 17:00.}

\item{dateTimeEnd}{a dateTime object that defines the end of the time period to be divided. If not specified, defaults to the latest `timestamp` in `edges`. Must be in one of the following formats: "YYYY-MM-DD hh:mm:ss" or "YYYY-MM-DD hh:mm" or "YYYY-MM-DD". Hours must use 24 hour time--e.g. 5:00 pm would be 17:00.}

\item{id1Col}{name of the column in `edges` containing the ID of the first individual in a dyad}

\item{id2Col}{name of the column in `edges` containing the ID of the second individual in a dyad}

\item{maxTimestampCol}{the name of the column to use for assigning the maximum timestamp, if no `dateTimeEnd` is provided. Default is "maxTimestamp".}

\item{minTimestampCol}{the name of the column to use for assigning the minimum timestamp, if no `dateTimeStart` is provided. Default is "minTimestamp".}
}
\value{
A list of data frames (i.e. "edgelists")
}
\description{
This function takes an edge list (returned by `vultureUtils::get*Edges()`) and slices it temporally. Returns a list of data frames (edge lists). Includes specifications for whether the time slices should start at the beginning or end, and what to do with any odd-length slice if/when the specified time interval doesn't divide evenly into the amount of time given.
}
