% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{spaceTimeGroups}
\alias{spaceTimeGroups}
\title{Group points by space and time}
\usage{
spaceTimeGroups(
  dataset,
  distThreshold,
  consecThreshold = 2,
  crsToSet = "WGS84",
  timestampCol = "timestamp",
  timeThreshold = "10 minutes",
  idCol = "trackId",
  latCol = "location_lat",
  longCol = "location_long",
  returnDist = TRUE,
  fillNA = FALSE
)
}
\arguments{
\item{dataset}{a data frame, filtered by speed etc, to use to create spatiotemporal groups.}

\item{distThreshold}{distance threshold at which to consider that two individuals are interacting (m).}

\item{consecThreshold}{Passed to vultureUtils::consecEdges. In how many consecutive time groups must the two individuals interact in order to be included? Default is 2.}

\item{crsToSet}{if `feedingSites` is a data frame, what CRS to pass to sf::st_set_crs() (NOT transform!). If `feedingSites` is already an sf object, `crsToSet` will be overridden by whatever the object's CRS is, unless it is NA.}

\item{timestampCol}{Passed to spatsoc::group_times. Name of date time column(s). either 1 POSIXct or 2 IDate and ITime. e.g.: 'datetime' or c('idate', 'itime')}

\item{timeThreshold}{Passed to spatsoc::group_times. Threshold for grouping times. e.g.: '2 hours', '10 minutes', etc. if not provided, times will be matched exactly. Note that provided threshold must be in the expected format: '## unit'.}

\item{idCol}{Name of the column containing individual ID's of the vultures.}

\item{latCol}{Name of the column containing latitude values}

\item{longCol}{Name of the column containing longitude values}

\item{returnDist}{Passed to spatsoc::edge_dist. Boolean indicating if the distance between individuals should be returned. If FALSE (default), only ID1, ID2 columns (and timegroup, splitBy columns if provided) are returned. If TRUE, another column "distance" is returned indicating the distance between ID1 and ID2. Default is TRUE.}

\item{fillNA}{Passed to spatsoc::edge_dist. Boolean indicating if NAs should be returned for individuals that were not within the threshold distance of any other. If TRUE, NAs are returned. If FALSE, only edges between individuals within the threshold distance are returned. Default is FALSE.}
}
\value{
an edge list (data frame)
}
\description{
This function takes a data frame of filtered, cleaned points and uses functions from `spatsoc` to group them by space and time. Then it uses vultureUtils::consecEdges to remove edges that don't occur in enough consecutive time groups.
}
