% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{spikyAltitudesFilter}
\alias{spikyAltitudesFilter}
\title{Spiky Altitudes Filter}
\usage{
spikyAltitudesFilter(dataset, idCol)
}
\arguments{
\item{dataset}{A dataset with columns: idCol}

\item{idCol}{The name of the column in the dataset containing vulture ID's. Defaults to "Nili_id" (assuming you have joined the Nili_ids from the who's who table).}
}
\value{
A dataset with outliers removed
}
\description{
This function takes in a dataset and removes points based on distance traveled between two points.
This is done by comparing speeds calculated from latitude and longitude values to known speed thresholds.
Additionally, the distance between points are compared at night to remove outliers.
}
