% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{spikySpeedsFilter}
\alias{spikySpeedsFilter}
\title{Spiky Speeds Filter}
\usage{
spikySpeedsFilter(dataset, idCol, longCol, latCol)
}
\arguments{
\item{dataset}{A dataset with columns: idCol, longCol, latCol}

\item{idCol}{The name of the column in the dataset containing vulture ID's. Defaults to "Nili_id" (assuming you have joined the Nili_ids from the who's who table).}

\item{longCol}{The name of the column in the dataset containing longitude values. Defaults to "location_long.1".}

\item{latCol}{The name of the column in the dataset containing latitude values. Defaults to "location_lat.1".}
}
\value{
A dataset with outliers removed
}
\description{
This function takes in a dataset and removes points based on distance traveled between two points.
This is done by comparing speeds calculated from latitude and longitude values to known speed thresholds.
Additionally, the distance between points are compared at night to remove outliers.
}
