% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{tempHeightSpeedFilter}
\alias{tempHeightSpeedFilter}
\title{Temperature, barometric height, and ground speed filter}
\usage{
tempHeightSpeedFilter(dataset)
}
\arguments{
\item{dataset}{A dataset with columns names: external_temperature, barometric_height, and ground_speed}
}
\value{
A dataset with outliers removed
}
\description{
This function takes in a dataset and removes outliers with columns: external_temperature, barometric_height, and
ground_speed equal to zero or NA.
}
