% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vultureUtils-package.R
\docType{package}
\name{vultureUtils-package}
\alias{vultureUtils}
\alias{vultureUtils-package}
\title{vultureUtils: A Package for Network Analysis of TAU/UCLA Griffon Vulture Data}
\description{
This package is designed to work with the TAU/UCLA joint NSF/BSF-supported dataset of GPS-tracked griffon vultures (_Gyps fulvus_). It automates data cleaning and the creation of social networks.
}
\keyword{internal}
