% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportingFunctions.R
\name{calcSRI}
\alias{calcSRI}
\title{Calculate SRI}
\usage{
calcSRI(dataset, edges, idCol = "Nili_id", timegroupCol = "timegroup")
}
\arguments{
\item{dataset}{the cleaned dataset.}

\item{edges}{edgelist created by spatsoc, with self edges and duplicate edges removed.}

\item{idCol}{character. Name of the column containing ID values.}

\item{timegroupCol}{character. Name of the column containing timegroup values.}
}
\value{
A data frame containing ID1, ID2, and SRI value.
}
\description{
Calculates SRI based on timegroup and individual occurrence information
}
