% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{downloadVultures}
\alias{downloadVultures}
\title{Data download}
\usage{
downloadVultures(
  loginObject,
  extraSensors = F,
  removeDup = T,
  dateTimeStartUTC = NULL,
  dateTimeEndUTC = NULL
)
}
\arguments{
\item{loginObject}{A Movebank login object, created by passing a username and password to move::movebankLogin}

\item{extraSensors}{Whether to include extra sensors. Defaults to FALSE.}

\item{removeDup}{Whether to remove duplicated timestamps. Defaults to TRUE.}

\item{dateTimeStartUTC}{a POSIXct object, in UTC. Will be converted to character assuming UTC.}

\item{dateTimeEndUTC}{a POSIXct object, in UTC. Will be converted to character assuming UTC.}
}
\value{
A movestack
}
\description{
Download vulture project data from the Israel vulture study Movebank repository, with some minor specifications. Note that you must specify your movebank credentials.
}
