% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getRoostingEdges}
\alias{getRoostingEdges}
\title{Create co-roosting edge list XXX START HERE}
\usage{
getRoostingEdges(
  cleanedInMask,
  roostPolygons,
  roostBuffer = 50,
  consecThreshold = 2
)
}
\arguments{
\item{cleanedInMask}{The cleaned GPS dataset to be used to create the edge list. This should be the output from `vultureUtils::cleanData()`.}

\item{roostPolygons}{Roost polygons. Must be an sf object with a CRS that matches the dataset CRS.}

\item{roostBuffer}{Number of meters to buffer the roost polygons by before intersecting them. Default is 50 m.}

\item{consecThreshold}{Minimal number of co-occurrences for considering a viable pair of interacting vultures (default is 2 consecutive time steps). Passed to `vultureUtils::spaceTimeGroups()`. Must be numeric.}
}
\value{
An edge list containing the following columns: `timegroup` gives the numeric index of the timegroup during which the interaction takes place. `minTimestamp` and `maxTimestamp` give the beginning and end times of that timegroup. `ID1` is the trackID of the first individual in this edge, and `ID2` is the trackID of the second individual in this edge.
}
\description{
Given a dataset of GPS points, a geographic mask, and some roost polygons, create an edge list.
}
