% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{gpsJamFilter}
\alias{gpsJamFilter}
\title{GPS jamming filter}
\usage{
gpsJamFilter(
  dataset,
  mask,
  longCol = "location_long.1",
  latCol = "location_lat.1",
  idCol = "Nili_id"
)
}
\arguments{
\item{dataset}{A dataset with columns: longCol, latCol, idCol}

\item{mask}{The object to use to mask the data. Must be an sf object.}

\item{longCol}{The name of the column in the dataset containing longitude values. Defaults to "location_long.1". Passed to `vultureUtils::maskData()`.}

\item{latCol}{The name of the column in the dataset containing latitude values. Defaults to "location_lat.1". Passed to `vultureUtils::maskData()`.}

\item{idCol}{The name of the column in the dataset containing vulture ID's. Defaults to "Nili_id" (assuming you have joined the Nili_ids from the who's who table).}
}
\description{
This function takes in a dataset and removes points that lie within mask, given that mask is an sf object containing
polygons where GPS points are shifted to after jamming.
}
