% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{makeGraph}
\alias{makeGraph}
\title{Make a single graph from an edge list.}
\usage{
makeGraph(
  edges,
  weighted = FALSE,
  id1Col = "ID1",
  id2Col = "ID2",
  vertices = NULL
)
}
\arguments{
\item{edges}{a data frame containing edges and their associated `timegroup`s. Returned from vultureUtils::get*Edges().}

\item{weighted}{whether or not the resulting graphs should have weights attached. Default is FALSE (unweighted graph).}

\item{id1Col}{name of the column in `edges` containing the ID of the first individual in a dyad. Default is "ID1".}

\item{id2Col}{name of the column in `edges` containing the ID of the second individual in a dyad. Default is "ID2".}

\item{vertices}{optional. Either NULL (default) or a data frame/vector with vertex names and optional metadata. If a vector is provided, it will be coerced to a data frame (see documentation for igraph::graph_from_data_frame).}
}
\value{
an igraph object. An undirected graph that is either weighted or unweighted, depending on whether `weighted` is T or F.
}
\description{
This is basically a wrapper of igraph::graph_from_data_frame. Takes into account the formatting returned from vultureUtils::get*Edges().
}
