% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{makeGraph}
\alias{makeGraph}
\title{Make a single graph from an edge list.}
\usage{
makeGraph(edges, weighted = FALSE, id1Col = "ID1", id2Col = "ID2")
}
\arguments{
\item{edges}{a data frame containing edges and their associated `timegroup`s. Returned from vultureUtils::get*Edges().}

\item{weighted}{whether or not the resulting graphs should have weights attached. Default is FALSE (unweighted graph).}

\item{id1Col}{name of the column in `edges` containing the ID of the first individual in a dyad. Default is "ID1".}

\item{id2Col}{name of the column in `edges` containing the ID of the second individual in a dyad. Default is "ID2".}
}
\value{
an igraph object. An undirected graph that is either weighted or unweighted, depending on whether `weighted` is T or F.
}
\description{
This is basically a wrapper of igraph::graph_from_data_frame. Takes into account the formatting returned from vultureUtils::get*Edges().
}
