% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{makeGraphsList}
\alias{makeGraphsList}
\title{Make a list of graphs}
\usage{
makeGraphsList(
  dataList,
  weighted = FALSE,
  id1Col = "ID1",
  id2Col = "ID2",
  allVertices = FALSE,
  allVerticesVec = NULL
)
}
\arguments{
\item{dataList}{a list of edge lists (data frames).}

\item{weighted}{whether or not the resulting graphs should have weights attached}

\item{id1Col}{name of the column containing the ID of the first individual in a dyad}

\item{id2Col}{name of the column containing the ID of the second individual in a dyad}

\item{allVertices}{whether to include all possible vertices in all graphs produced, even if they are not involved in any edges in the current time slice. Default is FALSE}

\item{allVerticesVec}{vector to use to define all possible vertices.}
}
\value{
A list of igraph graph objects
}
\description{
Creates a list of igraph graphs (and the data to go along with them) from a list of edgelists.
}
