% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{maskData}
\alias{maskData}
\title{Mask dataset}
\usage{
maskData(
  dataset,
  mask,
  longCol = "location_long",
  latCol = "location_lat",
  crs
)
}
\arguments{
\item{dataset}{a dataset to mask}

\item{mask}{an sf object to use as the mask.}

\item{longCol}{the name of the column in the dataset containing longitude values}

\item{latCol}{the name of the column in the dataset containing latitude values}

\item{crs}{(To be passed to `st_set_crs()`). One of (i) character: a string accepted by GDAL, (ii) integer, a valid EPSG value (numeric), or (iii) an object of class crs.}
}
\value{
A masked data set.
}
\description{
Given a mask and a dataset, apply the mask to the dataset and return only locations inside the mask.
}
