% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{removeUnnecessaryVars}
\alias{removeUnnecessaryVars}
\title{Remove unnecessary vars}
\usage{
removeUnnecessaryVars(dataset, addlVars = NULL, keepVars = NULL)
}
\arguments{
\item{dataset}{a dataset to remove variables from. Must be a data frame.}

\item{addlVars}{a character vector of additional variables to remove}

\item{keepVars}{a character vector of variables to keep that would otherwise be removed.}
}
\value{
A dataset, with variables removed
}
\description{
Remove variables we don't need, so the data is smaller
}
