% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vultureUtils-package.R
\docType{package}
\name{vultureUtils-package}
\alias{vultureUtils}
\alias{vultureUtils-package}
\title{vultureUtils: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
